[refactor] rearrange file structure
