+++ lib/core/helpers.js
@@ -16,6 +16,7 @@
   exports.root = path.join(path.dirname(module.filename), '../../');
   exports.lib = path.join(exports.root, 'lib');
   exports.core = path.join(exports.lib, 'core');
+  exports.execute = path.join(exports.lib, 'execute');
 
   // Simple progress tracker, will execute callback when .list is empty
   function ProgressTracker(callback) {
