+++ lib/core/monitorAbstract.js
@@ -8,6 +8,7 @@
 
   var util = require('util');
   var events = require('events');
+
   var helpers = require('./helpers.js');
 
   function MonitorAbstract(options, ready) {
