+++ lib/executables/daemon.js
